Refactored command folder structure, cleaned up problems being thrown
