+++ src/main/java/frc/robot/OI.java
@@ -6,6 +6,8 @@
 package frc.robot;
 
 import frc.robot.commands.*;
+import frc.robot.commands.test.*;
+import frc.robot.commands.util.*;
 import frc.robot.utils.FilteredJoystick;
 import frc.robot.utils.filters.*;
