+++ src/main/java/frc/robot/commands/DeployHatchCommandGroup.java
@@ -2,7 +2,6 @@ package frc.robot.commands;
 
 import frc.robot.Robot;
 import frc.robot.OI;
-import edu.wpi.first.wpilibj.command.Command;
 import edu.wpi.first.wpilibj.command.CommandGroup;
 
 /**
