+++ src/main/java/frc/robot/commands/IntakeCancelCommand.java
@@ -2,20 +2,18 @@ package frc.robot.commands;
 
 import frc.robot.Robot;
 import frc.robot.OI;
-import frc.robot.Ports;
-import frc.robot.TuningParams;
 import edu.wpi.first.wpilibj.command.Command;
 
 /**
  * A command group implementing the sequence necessary to pull cargo into the robot.
  */
 /*public class IntakeCancelCommand extends Command
-{   
+{
     private OI.Mode mode;
 
     /**
-     *   
-     *   
+     *
+     *
      **/
     /*public IntakeCancelCommand(OI.Mode mode)
     {
