+++ src/main/java/frc/robot/commands/IntakeCommandGroup.java
@@ -2,7 +2,6 @@ package frc.robot.commands;
 
 import frc.robot.Robot;
 import frc.robot.OI;
-import frc.robot.Ports;
 import frc.robot.TuningParams;
 import edu.wpi.first.wpilibj.command.CommandGroup;
 
@@ -10,12 +9,12 @@ import edu.wpi.first.wpilibj.command.CommandGroup;
  * A command group implementing the sequence necessary to pull cargo into the robot.
  */
 /*public class IntakeCommandGroup extends CommandGroup
-{   
+{
     private OI.Mode mode;
 
     /**
-     *   
-     *   
+     *
+     *
      **/
     /*public IntakeCommandGroup(OI.Mode mode)
     {
