+++ src/main/java/frc/robot/commands/IntakeWaitForCargoCommand.java
@@ -1,6 +1,5 @@
 package frc.robot.commands;
 
-import frc.robot.Ports;
 import frc.robot.Robot;
 import frc.robot.OI;
 import edu.wpi.first.wpilibj.command.Command;
@@ -9,13 +8,13 @@ import edu.wpi.first.wpilibj.command.Command;
  * A command which executes until cargo either detected in or removed from the intake area.
  */
 public class IntakeWaitForCargoCommand extends Command
-{   
+{
     private OI.Mode mode;
     private boolean isPresent;
     private boolean isFinished;
 
     /**
-     * 
+     *
      * @param mode - the operating mode in which this command must run.
      * @param isPresent - the command completes when the cargo intake sensor is in this state.
      *               For example, setting this to true, means that the command exits when cargo
@@ -29,7 +28,7 @@ public class IntakeWaitForCargoCommand extends Command
         this.isPresent  = isPresent;
         this.isFinished = true;
     }
-  
+
 
     // Called just before this Command runs the first time
     protected void initialize()
