+++ src/main/java/frc/robot/commands/test/TestCommandGroup.java
@@ -1,7 +1,10 @@
-package frc.robot.commands;
+package frc.robot.commands.test;
 
 import frc.robot.Robot;
 import frc.robot.OI;
+import frc.robot.commands.*;
+import frc.robot.commands.util.*;
+
 import edu.wpi.first.wpilibj.command.CommandGroup;
 
 /**
