+++ src/main/java/frc/robot/commands/test/TestDriveCommand.java
@@ -1,4 +1,4 @@
-package frc.robot.commands;
+package frc.robot.commands.test;
 
 import frc.robot.Robot;
 import frc.robot.OI;
