+++ src/main/java/frc/robot/commands/util/ModeSelect.java
@@ -1,38 +1,50 @@
-package frc.robot.commands;
+package frc.robot.commands.util;
 
+import frc.robot.Robot;
+import frc.robot.OI;
 import edu.wpi.first.wpilibj.command.Command;
 
 /**
- * A class supporting timed delays. This is intended for use in
- * command groups where a delay is required between adjacent commands.
+ * A class sets the normal or override operator control mode
  */
-public class DelayCommand extends Command
+public class ModeSelect extends Command
 {
-    private int msDelay;
-    private int msToGo;
-    private final int msPerPeriodic = 20;
 
-    public DelayCommand(int msDelay)
+    /**
+     *
+     * @param Override sets manual override mode if true, else sets normal control mode.
+     */
+    public ModeSelect()
     {
-        this.msDelay = msDelay;
+
     }
+
 
     // Called just before this Command runs the first time
     protected void initialize()
     {
-        msToGo = msDelay;
+        OI.Mode RobotState = Robot.oi.getMode();
+        if (RobotState == OI.Mode.NORMAL)
+        {
+            Robot.oi.setMode(OI.Mode.MANUAL);
+        }
+        else if(RobotState == OI.Mode.MANUAL)
+        {
+            Robot.oi.setMode(OI.Mode.NORMAL);
+        }
     }
 
     // Called repeatedly when this Command is scheduled to run
     protected void execute()
     {
-        msToGo -= msPerPeriodic;
+
+
     }
 
     // Make this return true when this Command no longer needs to run execute()
     protected boolean isFinished()
     {
-        return (msToGo <= 0);
+        return true;
     }
 
     // Called once after isFinished returns true
