+++ src/main/java/frc/robot/commands/util/NullCommand.java
@@ -1,44 +1,33 @@
-package frc.robot.commands;
+package frc.robot.commands.util;
 
-import frc.robot.Robot;
-import frc.robot.OI;
 import edu.wpi.first.wpilibj.command.Command;
 
 /**
- * A class sets the normal or override operator control mode
+ * A class which implements a command that does absolutely nothing. This
+ * may sound silly but it's useful when switching modes and making some
+ * buttons inactive.
  */
-public class ModeSelect extends Command
+public class NullCommand extends Command
 {
 
     /**
-     * 
-     * @param Override sets manual override mode if true, else sets normal control mode.
+     *
+     * @param
      */
-    public ModeSelect()
+    public NullCommand()
     {
-        
     }
-  
+
 
     // Called just before this Command runs the first time
     protected void initialize()
     {
-        OI.Mode RobotState = Robot.oi.getMode();
-        if (RobotState == OI.Mode.NORMAL)
-        {
-            Robot.oi.setMode(OI.Mode.MANUAL);
-        }
-        else if(RobotState == OI.Mode.MANUAL)
-        {
-            Robot.oi.setMode(OI.Mode.NORMAL);
-        }
     }
 
     // Called repeatedly when this Command is scheduled to run
     protected void execute()
     {
-        
-        
+
     }
 
     // Make this return true when this Command no longer needs to run execute()
